services/intraday-prediction-service/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
pydantic>=2.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
lightgbm>=4.3
sqlalchemy>=2.0
psycopg2-binary>=2.9
prometheus-client>=0.20
